podział na poszczegolne pliki, dodanie obslugi zawnetrznych funkcji utils, modyfikacja metody sprawdz
